add slice
