+++ project/src/components/error-message/error-message.tsx
@@ -2,7 +2,7 @@ import { useAppSelector } from '../../hooks';
 import './error-message.css';
 
 const ErrorMessage = () => {
-  const error = useAppSelector((state) => state.reducer.error);
+  const error = useAppSelector((state) => state.OFFER.error);
 
   return (error)
     ? <div className='error-message'>{error}</div>
