+++ project/src/store/comments-process/types.ts
@@ -1,0 +1,7 @@
+import { Comment } from '../../types/comments';
+import { Review } from '../../types/review';
+
+export type CommentProcess = {
+  loadComments: Comment[];
+  nextReview: Review | null;
+}
