+++ project/src/store/near-offers-process/types.ts
@@ -1,0 +1,5 @@
+import { Offers } from '../../types/offers';
+
+export type NearOffersProcess = {
+  nearOffers: Offers[];
+}
