+++ project/src/store/user-process/types.ts
@@ -1,3 +1,6 @@
+import { AuthorizationStatus } from '../../constants/constants';
+import { AppRoute } from '../../router/RoutePath';
+
 export type UserData = {
   id: number;
   email: string;
@@ -6,3 +9,9 @@ export type UserData = {
   isPro: boolean;
   avatarUrl: string;
 } | null;
+
+export type UserProcess = {
+  authorizationStatus: AuthorizationStatus;
+  userData: UserData;
+  redirectToRoute: AppRoute | null;
+};
